agent_local/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pyyaml>=6.0
requests>=2.31.0
